backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
aiofiles>=24.1.0

sqlalchemy>=2.0.0
alembic>=1.13.0

pydantic>=2.8.0
pydantic-settings>=2.4.0

python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

Pillow>=10.4.0
pillow-heif>=0.18.0
opencv-python-headless>=4.10.0

numpy>=1.26.0
scikit-learn>=1.5.0
timm>=1.0.0
scipy>=1.10.0
tqdm>=4.66.0

torch>=2.5.0
torchvision>=0.20.0
onnx>=1.16.0
# onnxruntime-gpu>=1.18.0  # Uncomment strictly for GPU environments

matplotlib>=3.9.0
seaborn>=0.13.0

optuna>=3.6.0
tensorboard>=2.17.0
